fix: rename version.h to myathan_version.h (avoid C++ stdlib conflict)
<version> is a C++20 standard library header. Our version.h was
conflicting during PlatformIO builds. Renamed to myathan_version.h
and updated all includes.
